Reshuffle files and enforce `from jax._src.random import KeyArray`
