+++ lib/param_utils/init_params.py
@@ -1,3 +1,4 @@
+from jax._src.random import KeyArray
 import jax.random as rand
 from transformers import BartConfig, FlaxBartModel
 
@@ -6,7 +7,7 @@ from ..random.wrapper import key2seed
 
 config = BartConfig.from_pretrained('facebook/bart-base')
 
-def init_params(key: rand.KeyArray) -> dict:
+def init_params(key: KeyArray) -> dict:
     seed = key2seed(key)
     model_flax = FlaxBartModel(config=config, seed=seed)
     model_jax = flax2jax(model_flax.params)
